Rename of bundles related methods arguments to avoid confusion with data structures.
Bundles components are called "elements" in the specification.
Methods such as send_bundle, sync, clump_bundle and calc_bndl_dgram_size
used the word bundle to refer to these "elements", a list of lists
without timetag. That creates a bit of confusion with conceptual
bundle format, a list starting with a timetag and then the "elements".

We don't use the word bundle to refer to the elements to avoid
that confusion. Elements are lists of lists without timetag.

Also the order of the sync methods parameters, latency (time to
timtag) and elements was inverted for a similar reason.