A possible event stream cleanup solution.
Resources are either managed from EventStreamPlayer
or individually if they are used without a thread
player (manually calling next). In the later case
possible sub event streams will not be able to free
their resources if they don't finish.

The other solution, that I'm avoiding by now, is
to add a general purpose finalizer to TimeThread,
but I'm not sure that the second case above has
a solution anyway.