Restore TempoClock _base_seconds to current_tt._seconds.
Sclang calls update_logical_time ONLY once when evaluating
a block of code. Here that is not possible. In this case
the clock beats are updated once (as in sclang) if and
only if the code is executed within a routine. If they are
evaluated individually, the beats of the clocks are left
with the offset to which they were initialized, and that
looks weird but is correct.