Add play function (experimental)
In this case is also a shortcut to play events from dictionaries
or keyword arguments which seems to be more important because we
don't have a syntax shortcut. The problem is that there is no way
to play a function with server arguments because of that. This
may change in the future, somehow.