Routines' seconds and associated clock are both internal.
The implementation is the same only this members are considered internal.
SystemClock and TempoClock both represent current logical time inside running
routines, and routines keep the same time reference when scheduled. Logical
time is coupled to routines but a routine can change the clock when resumed
after a pause, and the same clock can schedule many different routines.
Routines' time properties could be used to know at which beat it was stopped
but will always update logical time when resumed as if the time was always
running (because logical time runs relative to a physical time start point).
Historically, the documentation explains the internals rather than what is
observed by the user and that may misleading about its usage, e.g. why routines
don't start always from beat zero? Because they don't keep their own internal
state as a timeline and is not how much time it was running itself but at what
clock's beats, the clock would be the timeline in this case.